# docs/swagger.yaml
basePath: /
definitions:
  handlers.AddCartItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: E-commerce REST API example
  title: Golang REST API
  version: 0.1.0
paths:
  /cart:
    post:
      consumes:
      - application/json
      description: Create a new cart for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a cart
      tags:
      - Cart
  /cart/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete a user's cart (admin only)
      parameters:
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete cart by user ID (Admin)
      tags:
      - Cart
    get:
      consumes:
      - application/json
      description: Retrieve a user's cart (admin only)
      parameters:
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get cart by user ID (Admin)
      tags:
      - Cart
  /cart/item:
    post:
      consumes:
      - application/json
      description: Add a product to cart or update quantity
      parameters:
      - description: Cart item data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/handlers.AddCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add or update cart item
      tags:
      - Cart
  /cart/item/{product_id}:
    delete:
      consumes:
      - application/json
      description: Remove a specific product from cart
      parameters:
      - description: Product UUID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Remove item from cart
      tags:
      - Cart
  /cart/items:
    get:
      consumes:
      - application/json
      description: Retrieve all items in user's cart (uses cache)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all cart items
      tags:
      - Cart
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"